Return
Retomada de mudanças

+++ footer.php
@@ -1,3 +1,17 @@
+	<footer class="d-none d-md-block row">
+	  <section class="col">
+        <p class="m-0 py-1"> 
+        <?php 
+        echo "Versão <b>".ERP_VERSION."</b> do sistema";
+	  	$filename = ABSPATH."global_config.php";
+	  	if(file_exists($filename)){
+	  	  echo ", última atualização em <b>" . date ("d/m/Y H:i:s", filemtime($filename))."</b>";
+	  	}
+	  	?>
+	  	</p>
+	  </section>
+	</footer>
+
 	<!-- jQuery 3.2.1 -->
 	<script type="text/javascript" src="js/jquery-3.2.1.min.js"></script>
 
@@ -29,12 +43,14 @@
     // Load de scripts do módulo carregado
     if(isset($_GET["mod"])){
       $modulo = filter_input(INPUT_GET, "mod", FILTER_SANITIZE_SPECIAL_CHARS);
-      $js_files = scandir( MODULES_ABSPATH."/{$modulo}/js" );
-      foreach( $js_files as $file ){
-        if( is_file( MODULES_ABSPATH."/{$modulo}/js/{$file}" )){
-          echo "<script type='text/javascript' src='modules/{$modulo}/js/{$file}'></script>";
-        }
-      }
+	  if(file_exists(MODULES_ABSPATH."/{$modulo}/js")){ 
+        $js_files = scandir( MODULES_ABSPATH."/{$modulo}/js" );
+	    foreach( $js_files as $file ){
+	      if( is_file( MODULES_ABSPATH."/{$modulo}/js/{$file}" )){
+	        echo "<script type='text/javascript' src='modules/{$modulo}/js/{$file}'></script>";
+	      }
+	    }
+	  }
     }
 	?>
   </body>
